Move MakeTestNodeConfig to utils. Part of #665
